+++ CHANGELOG.md
@@ -109,6 +109,7 @@
 - `Polyline.Intersects(Polygon polygon, out List<Polyline> sharedSegments)` bug when polyline start/end is on polygon perimeter
 - `GltfBufferExtensions.CombineBufferAndFixRefs` bug when combining buffers from multiple gltf files.
 - `Obstacle.FromWall` was failing when producing a polygon.
+- `Network` incorrect tree building and search of intersections
 
 ## 1.1.0